extract loading state into its own component
