+++ apps/web/src/features/spaces/components/AuthState/index.tsx
@@ -2,19 +2,10 @@ import type { ReactNode } from 'react'
 import SignedOutState from '@/features/spaces/components/SignedOutState'
 import { isUnauthorized } from '@/features/spaces/utils'
 import UnauthorizedState from '@/features/spaces/components/UnauthorizedState'
+import LoadingState from '@/features/spaces/components/LoadingState'
 import { useAppSelector } from '@/store'
 import { isAuthenticated } from '@/store/authSlice'
 import { useOrganizationsGetOneV1Query } from '@safe-global/store/gateway/AUTO_GENERATED/organizations'
-import { Box, CircularProgress } from '@mui/material'
-
-// Loading state component with centered spinner
-export const LoadingState = () => {
-  return (
-    <Box display="flex" justifyContent="center" alignItems="center" minHeight="100vh">
-      <CircularProgress aria-label="Loading content" />
-    </Box>
-  )
-}
 
 const AuthState = ({ spaceId, children }: { spaceId: string; children: ReactNode }) => {
   const isUserSignedIn = useAppSelector(isAuthenticated)
