+++ apps/web/src/features/spaces/components/LoadingState/index.tsx
@@ -1,0 +1,11 @@
+import { Box, CircularProgress } from '@mui/material'
+
+const LoadingState = () => {
+  return (
+    <Box display="flex" justifyContent="center" alignItems="center" minHeight="100vh">
+      <CircularProgress aria-label="Loading content" />
+    </Box>
+  )
+}
+
+export default LoadingState
